+++ docs/CHANGELOG.md
@@ -16,10 +16,12 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 ### Changed
 * Rename FPSCounter, FPSCounterBundle, FPSCounterSystem to FpsCounter, FpsCounterBundle, FpsCounterSystem. ([#1719])
 * Add Tint component support for sprites. ([#1756])
+* Remove remaining <N: RealField> type parameter on GameDataBuilder, add Debug derive to LoggerConfig ([#1758])
 
 [#1719]: https://github.com/amethyst/amethyst/pull/1719
 [#1756]: https://github.com/amethyst/amethyst/pull/1756
 [#1733]: https://github.com/amethyst/amethyst/pull/1733
+[#1758]: https://github.com/amethyst/amethyst/pull/1758
 
 ## [0.11.0] - 2019-06